requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
idna==2.10
mysql==0.0.2
mysql-connector==2.2.9
mysqlclient==2.0.1
numpy==1.19.2
pandas==1.1.3
psycopg2-binary==2.8.6
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.10
